Consolidated duplicate code in queue store implementations. Removed an instance of needing to check datastore ID to run specific datastore code in server.go
